Split up glTF backend into separate modules
